+++ Appwrite/Enums/BackupServices.cs
@@ -12,6 +12,9 @@ namespace Appwrite.Enums
         }
 
         public static BackupServices Databases => new BackupServices("databases");
+        public static BackupServices Tablesdb => new BackupServices("tablesdb");
+        public static BackupServices Documentsdb => new BackupServices("documentsdb");
+        public static BackupServices Vectorsdb => new BackupServices("vectorsdb");
         public static BackupServices Functions => new BackupServices("functions");
         public static BackupServices Storage => new BackupServices("storage");
     }