+++ Appwrite/Services/Functions.cs
@@ -54,7 +54,7 @@ namespace Appwrite.Services
         /// API.
         /// </para>
         /// </summary>
-        public Task<Models.Function> Create(string functionId, string name, Appwrite.Enums.Runtime runtime, List<string>? execute = null, List<string>? events = null, string? schedule = null, long? timeout = null, bool? enabled = null, bool? logging = null, string? entrypoint = null, string? commands = null, List<Appwrite.Enums.Scopes>? scopes = null, string? installationId = null, string? providerRepositoryId = null, string? providerBranch = null, bool? providerSilentMode = null, string? providerRootDirectory = null, string? specification = null)
+        public Task<Models.Function> Create(string functionId, string name, Appwrite.Enums.Runtime runtime, List<string>? execute = null, List<string>? events = null, string? schedule = null, long? timeout = null, bool? enabled = null, bool? logging = null, string? entrypoint = null, string? commands = null, List<Appwrite.Enums.Scopes>? scopes = null, string? installationId = null, string? providerRepositoryId = null, string? providerBranch = null, bool? providerSilentMode = null, string? providerRootDirectory = null, string? buildSpecification = null, string? runtimeSpecification = null, long? deploymentRetention = null)
         {
             var apiPath = "/functions";
 
@@ -77,7 +77,9 @@ namespace Appwrite.Services
                 { "providerBranch", providerBranch },
                 { "providerSilentMode", providerSilentMode },
                 { "providerRootDirectory", providerRootDirectory },
-                { "specification", specification }
+                { "buildSpecification", buildSpecification },
+                { "runtimeSpecification", runtimeSpecification },
+                { "deploymentRetention", deploymentRetention }
             };
 
             var apiHeaders = new Dictionary<string, string>()
@@ -190,7 +192,7 @@ namespace Appwrite.Services
         /// Update function by its unique ID.
         /// </para>
         /// </summary>
-        public Task<Models.Function> Update(string functionId, string name, Appwrite.Enums.Runtime? runtime = null, List<string>? execute = null, List<string>? events = null, string? schedule = null, long? timeout = null, bool? enabled = null, bool? logging = null, string? entrypoint = null, string? commands = null, List<Appwrite.Enums.Scopes>? scopes = null, string? installationId = null, string? providerRepositoryId = null, string? providerBranch = null, bool? providerSilentMode = null, string? providerRootDirectory = null, string? specification = null)
+        public Task<Models.Function> Update(string functionId, string name, Appwrite.Enums.Runtime? runtime = null, List<string>? execute = null, List<string>? events = null, string? schedule = null, long? timeout = null, bool? enabled = null, bool? logging = null, string? entrypoint = null, string? commands = null, List<Appwrite.Enums.Scopes>? scopes = null, string? installationId = null, string? providerRepositoryId = null, string? providerBranch = null, bool? providerSilentMode = null, string? providerRootDirectory = null, string? buildSpecification = null, string? runtimeSpecification = null, long? deploymentRetention = null)
         {
             var apiPath = "/functions/{functionId}"
                 .Replace("{functionId}", functionId);
@@ -213,7 +215,9 @@ namespace Appwrite.Services
                 { "providerBranch", providerBranch },
                 { "providerSilentMode", providerSilentMode },
                 { "providerRootDirectory", providerRootDirectory },
-                { "specification", specification }
+                { "buildSpecification", buildSpecification },
+                { "runtimeSpecification", runtimeSpecification },
+                { "deploymentRetention", deploymentRetention }
             };
 
             var apiHeaders = new Dictionary<string, string>()